Move custom array types to testhelpers and use TSlow in offsetarray/collect_to test.
